Memoize ChatMessage, share entity tokenization and pill rendering, gate image bubbles on visibility
