+++ apps/web-app/src/ErrorBoundary.tsx
@@ -2,7 +2,6 @@ import React, { type ReactNode } from "react";
 
 interface ErrorBoundaryProps {
   children: ReactNode;
-  fallback?: ReactNode;
 }
 
 interface ErrorBoundaryState {
