--- apps/web-app/src/LinkyLogo.tsx
@@ -1,55 +1,0 @@
-import React from "react";
-
-type Props = {
-  className?: string;
-  title?: string;
-};
-
-const LinkyLogo: React.FC<Props> = ({ className, title }) => {
-  return (
-    <svg
-      className={className}
-      viewBox="0 0 800 800"
-      role={title ? "img" : "presentation"}
-      aria-label={title}
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <defs>
-        <linearGradient id="linkyGradient" x1="0" y1="0" x2="1" y2="1">
-          <stop offset="0" stopColor="#14b8a6" />
-          <stop offset="1" stopColor="#0d9488" />
-        </linearGradient>
-      </defs>
-
-      <path
-        fill="url(#linkyGradient)"
-        opacity="1"
-        stroke="none"
-        d="M207.712494,447.802002 C199.531906,450.803680 191.660492,453.516113 184.123413,457.094421 C182.611099,457.812439 180.668594,459.669800 179.335037,457.402588 C178.155502,455.397247 180.354721,454.141937 181.652542,452.975586 C190.235779,445.261688 200.613327,440.697937 211.163574,436.536987 C241.898087,424.415436 273.819489,416.142120 305.685974,407.729828 C322.747040,403.225922 339.776733,398.646637 356.601593,393.307007 C360.277924,392.140228 362.218903,389.976959 363.914734,386.582672 C383.575714,347.230469 403.578033,308.026917 430.788361,273.163757 C447.240326,252.084641 467.107330,235.127640 492.856262,226.173630 C501.701843,223.097641 510.738190,222.300644 519.845886,225.003220 C529.560608,227.885895 533.930786,235.258652 535.242859,244.603409 C537.091980,257.772583 533.265381,270.012482 527.825256,281.746002 C513.376465,312.909515 490.426758,336.982574 463.584351,357.660553 C441.553650,374.631927 417.659882,388.380829 391.553101,398.134705 C389.454712,398.918701 387.642670,399.846375 386.556458,401.953552 C373.780212,426.738251 358.930054,450.220398 342.189331,472.514923 C341.507141,473.423492 341.065369,474.512512 340.307159,475.888519 C345.726379,479.542633 351.980377,480.823395 357.748169,483.106110 C388.097015,495.117462 418.816010,506.132202 450.520142,514.005981 C467.477753,518.217468 484.850800,520.271240 502.292786,521.792419 C519.178467,523.265137 535.907776,522.448120 552.524353,519.633484 C574.728271,515.872498 595.452820,508.644714 611.463013,491.732788 C613.727173,489.341095 615.734375,486.557373 617.261292,483.641357 C625.657959,467.605896 617.186279,453.271118 598.862427,452.348145 C595.912598,452.199554 592.948914,452.327301 590.178467,452.327301 C589.746826,450.671875 590.360046,450.102295 591.170166,449.813568 C605.730347,444.624847 620.141296,443.335175 633.772339,452.365753 C646.708374,460.935883 651.370117,474.103760 646.794617,489.471863 C640.751221,509.770325 626.851196,523.816345 609.257141,534.415222 C577.316650,553.656494 542.177368,558.260559 505.732117,555.803833 C456.903625,552.512451 412.078247,535.147522 367.782715,516.031555 C353.717163,509.961548 339.606140,503.994141 325.605621,497.777832 C322.188354,496.260529 319.964661,496.838776 317.236908,499.340515 C292.065369,522.426575 263.189758,538.772888 229.530457,545.639465 C212.125122,549.190247 194.573944,550.020386 177.198410,544.478088 C147.063492,534.866150 144.893295,504.962097 162.310562,485.011017 C175.473236,469.933441 193.190018,462.809479 212.266525,458.741058 C240.918228,452.630615 269.199554,455.712219 297.284180,462.620850 C302.268158,463.846893 307.942902,467.045563 312.062988,465.659729 C316.190155,464.271484 318.887421,458.373688 322.006317,454.295624 C330.718109,442.904877 338.401917,430.816833 345.838409,418.571533 C346.676727,417.191162 347.837280,415.871582 347.665833,412.796417 C301.172546,425.317413 253.516068,431.883087 207.712494,447.802002 M471.647705,259.148346 C467.695343,263.648712 463.508942,267.966339 459.828766,272.679260 C436.601532,302.425079 420.307434,336.215149 403.567566,369.732269 C403.003723,370.861267 401.711700,372.247620 403.015198,373.827179 C404.995239,373.917664 406.572998,372.894196 408.211700,372.127411 C423.517517,364.965515 437.412781,355.500977 451.062134,345.666382 C476.931641,327.027008 498.748871,304.650543 514.038513,276.401520 C518.422119,268.302429 521.380127,259.631622 522.566223,250.458939 C523.739136,241.388443 517.962341,234.596451 509.707764,235.481369 C505.422272,235.940796 501.263550,237.140732 497.346893,239.186066 C487.776917,244.183670 479.656525,251.006088 471.647705,259.148346 M176.029984,528.589050 C184.672729,532.121460 193.728043,533.455139 202.929413,532.495178 C236.999344,528.940613 265.010254,512.937317 289.406250,489.735657 C290.554932,488.643158 292.168549,487.760590 292.169800,485.488495 C289.918030,483.626495 287.060730,482.813538 284.193512,482.014862 C271.523987,478.485870 258.783875,475.427643 245.581482,474.539581 C224.921463,473.149872 205.018799,475.464172 186.739426,486.006958 C179.746170,490.040375 173.685211,495.138000 169.652145,502.368011 C163.679337,513.075439 165.555725,521.837769 176.029984,528.589050 z"
-      />
-
-      <path
-        fill="url(#linkyGradient)"
-        opacity="1"
-        stroke="none"
-        d="M656.840454,347.349152 C657.892334,350.210114 658.990295,352.658569 659.678528,355.217285 C662.796021,366.806702 654.756104,376.215363 642.160461,375.683777 C639.014343,375.551025 635.890259,374.934387 632.751221,374.583496 C630.738098,374.358521 628.772095,374.483215 626.576050,376.035583 C628.922791,380.608643 633.451050,382.075317 637.663635,383.536224 C643.786194,385.659515 650.122192,387.166107 656.056946,389.853394 C665.872620,394.297974 672.418640,401.656281 675.486389,411.981262 C676.041626,413.850159 676.164307,415.942719 676.024170,417.898865 C675.851501,420.307465 674.330688,421.917480 671.820251,422.052490 C669.409912,422.182037 667.441956,421.103149 667.001831,418.614624 C663.864319,400.872681 649.548279,397.196838 635.363892,393.354004 C633.019226,392.718781 631.215576,393.871216 629.519287,395.391968 C608.016174,414.670380 578.630066,409.944000 562.115540,394.729919 C560.821716,393.537994 559.627625,392.574646 557.725769,393.245789 C550.685181,395.730377 543.326294,397.479492 537.119934,401.910767 C531.640442,405.823029 527.812988,410.955170 526.429871,417.635773 C525.774048,420.803162 524.410706,423.214294 520.870911,422.579742 C517.101990,421.904083 516.463257,418.790436 516.992371,415.593140 C518.913269,403.985779 525.879333,396.130096 536.045349,390.902008 C542.126770,387.774506 548.667847,385.765686 555.152771,383.623535 C559.202271,382.285858 563.230896,380.672394 565.734314,376.755646 C563.965454,374.412720 561.809021,374.760010 559.667114,374.966431 C555.862244,375.333191 552.017456,376.176056 548.252930,375.912384 C534.318298,374.936615 528.294861,363.534241 534.655640,349.679199 C539.074097,340.054718 542.425476,330.254578 543.447327,319.686279 C544.182739,312.079590 544.612854,304.406036 546.761169,297.020355 C553.002502,275.563782 564.279846,259.017914 588.075256,254.509659 C609.829529,250.388123 629.757507,260.589630 639.558228,280.723785 C645.501770,292.934052 647.786438,306.006897 649.180969,319.245758 C650.205322,328.971039 652.464233,338.207336 656.840454,347.349152 M628.785217,314.808105 C626.999268,313.014160 624.959412,311.549194 622.551758,310.785278 C609.832520,306.749725 598.572388,300.281250 588.764343,291.185333 C584.854370,287.559296 583.559143,287.888885 581.538025,292.626678 C578.143250,300.584503 573.479431,307.514801 566.399963,312.658722 C562.099426,315.783569 560.401245,319.981903 561.042114,325.258026 C562.790894,339.656403 570.444641,350.707397 581.295959,359.550995 C592.190857,368.430176 601.384033,368.099518 612.353882,359.227631 C620.158447,352.915619 625.751282,344.849579 629.155151,335.483521 C631.531860,328.943909 633.763062,322.094910 628.785217,314.808105 M609.619934,371.389801 C601.152283,375.287903 592.628601,375.955658 584.153198,371.536133 C581.181091,369.986389 579.589722,370.630676 578.634460,373.817841 C577.314697,378.221100 574.877808,382.030029 571.319946,384.924286 C568.648315,387.097595 569.375610,388.641907 571.560242,390.482880 C577.096741,395.148376 583.552368,397.651306 590.643311,398.432526 C603.486572,399.847534 614.974487,397.022949 624.770935,387.022247 C619.110168,383.198120 615.170105,378.640625 613.785828,372.279144 C613.318848,370.133301 611.842163,370.382019 609.619934,371.389801 z"
-      />
-
-      <path
-        fill="url(#linkyGradient)"
-        opacity="1"
-        stroke="none"
-        d="M223.504364,358.951385 C209.219421,353.234833 200.110474,342.774414 193.689438,329.674316 C192.623688,327.500000 191.423691,325.916779 189.236908,324.660065 C181.176651,320.027863 177.416336,307.456085 181.950424,299.321106 C183.734543,296.120148 183.539017,293.567474 182.534073,290.417908 C179.843842,281.986511 177.902496,273.512482 179.604889,264.464966 C181.428604,254.772690 187.741684,247.756149 197.434219,246.231781 C200.447968,245.757812 202.577332,244.361664 204.731339,242.414429 C217.477737,230.891464 242.345688,229.162140 255.899780,239.644379 C262.447113,244.707855 267.902588,250.958008 275.589478,254.834198 C280.735138,257.428925 281.936005,263.524506 282.322021,269.170197 C282.837128,276.703400 282.001617,284.113892 279.858704,291.363159 C279.231873,293.483582 279.036865,295.397217 280.249115,297.557220 C285.279114,306.519714 282.728302,318.347107 274.372589,324.213226 C271.794220,326.023346 270.128021,328.166962 268.715210,330.890076 C263.251526,341.421051 255.733170,350.172363 245.543350,356.434143 C238.753738,360.606476 231.461823,361.076874 223.504364,358.951385 M272.282043,314.587830 C273.735931,310.828247 276.038971,305.614990 272.257111,303.605835 C267.211975,300.925598 267.467651,297.048462 266.816864,293.023743 C264.922546,281.308533 261.007996,278.121918 249.197983,279.497803 C235.287659,281.118408 221.966415,279.231445 209.306229,273.261505 C206.569138,271.970856 205.205063,272.756531 203.517960,274.939331 C198.548050,281.369476 196.125137,288.691742 195.480576,296.642761 C195.238907,299.624023 195.202240,302.116608 191.356140,302.848785 C189.449341,303.211731 188.942093,305.548615 189.172333,307.455566 C189.720917,311.999054 190.341919,316.610992 196.017670,317.760651 C198.281693,318.219208 198.578995,320.176056 199.247299,321.816925 C202.960434,330.933746 208.479980,338.726257 215.981613,345.134216 C226.102173,353.779266 236.934982,354.038269 246.726715,345.668884 C251.695297,341.421967 256.078186,336.566986 259.366180,330.930695 C262.791138,325.059601 264.821777,318.190674 272.282043,314.587830 z"
-      />
-
-      <path
-        fill="url(#linkyGradient)"
-        opacity="1"
-        stroke="none"
-        d="M204.563980,377.615173 C203.704056,377.154144 202.958939,376.571838 202.506622,376.746765 C190.746277,381.294495 178.237442,383.914856 167.407471,390.871460 C158.181107,396.798004 153.102036,405.012207 152.336655,415.897949 C152.112167,419.090729 151.055634,421.603027 147.469269,421.607452 C143.496048,421.612335 142.742935,418.843048 142.664154,415.456757 C142.301025,399.851776 152.086899,385.600983 168.245911,378.935089 C176.079926,375.703461 184.142303,373.028625 192.072113,370.024689 C196.750748,368.252380 201.218582,366.224670 204.740128,362.347382 C207.249023,359.585052 210.831924,360.874268 210.971695,364.756378 C211.135681,369.310669 212.414902,372.736847 216.257187,375.320038 C221.262482,378.685120 226.820435,380.488159 232.696304,380.281311 C237.129120,380.125244 241.629288,378.912598 245.447800,376.355804 C249.074768,373.927216 252.494263,371.308594 252.041275,366.056763 C251.874939,364.128204 252.407471,361.912689 254.597549,361.039246 C256.831512,360.148346 258.348358,361.516022 259.750000,363.091522 C264.784943,368.750916 271.924713,370.337219 278.641113,372.683136 C284.454163,374.713501 290.309937,376.633179 295.816284,379.413849 C300.473877,381.765869 304.635712,384.754974 307.430237,389.339996 C308.828125,391.633514 309.869263,394.084412 307.379120,396.049347 C304.772827,398.105988 301.331024,398.363129 299.240204,395.728668 C289.758606,383.781769 274.693329,383.249298 262.349976,377.199066 C259.894409,375.995453 258.743988,378.040131 257.264832,379.251770 C240.584412,392.914764 220.754044,392.384552 204.563980,377.615173 z"
-      />
-    </svg>
-  );
-};
-
-export default LinkyLogo;
