+++ apps/web-app/src/components/ProfileAvatarEditor.tsx
@@ -1,10 +1,7 @@
 import React from "react";
-import {
-  AVATAR_EDITOR_CONTROLS,
-  type AvatarEditorControlId,
-} from "../derivedProfile";
+import type { AvatarEditorControlId } from "../derivedProfile";
 import { formatShortNpub, getInitials } from "../utils/formatting";
-import { AvatarEditorIcon } from "./AvatarEditorIcon";
+import { AvatarControlGrid } from "./AvatarControlGrid";
 import { AvatarPhotoInput } from "./AvatarPhotoInput";
 
 interface ProfileAvatarEditorProps {
@@ -73,52 +70,13 @@ export function ProfileAvatarEditor({
         t={t}
       />
 
-      <div
-        className="onboarding-avatar-grid"
-        role="list"
-        aria-label={t("onboardingAvatarGridLabel")}
-      >
-        {AVATAR_EDITOR_CONTROLS.map((control) => (
-          <button
-            key={control.id}
-            type="button"
-            className="onboarding-avatar-choice onboarding-avatar-editButton"
-            onClick={() => cycleProfileAvatarControl(control.id)}
-            aria-label={control.label}
-            title={control.label}
-          >
-            <span
-              className="onboarding-avatar-choicePlus onboarding-avatar-editEmoji"
-              aria-hidden="true"
-            >
-              <AvatarEditorIcon controlId={control.id} />
-            </span>
-          </button>
-        ))}
-
-        <button
-          type="button"
-          className={`onboarding-avatar-choice onboarding-avatar-choiceCustom${profileSelectedPictureKind === "custom" ? " is-selected" : ""}`}
-          onClick={() => void onPickProfilePhoto()}
-          aria-pressed={profileSelectedPictureKind === "custom"}
-        >
-          <span className="onboarding-avatar-choicePlus" aria-hidden="true">
-            {profileCustomPictureUrl ? (
-              <img
-                src={profileCustomPictureUrl}
-                alt=""
-                loading="lazy"
-                referrerPolicy="no-referrer"
-              />
-            ) : (
-              "+"
-            )}
-          </span>
-          <span className="onboarding-avatar-choiceLabel">
-            {t("profileUploadPhoto")}
-          </span>
-        </button>
-      </div>
+      <AvatarControlGrid
+        customPictureUrl={profileCustomPictureUrl}
+        isCustomSelected={profileSelectedPictureKind === "custom"}
+        onCycle={cycleProfileAvatarControl}
+        onPickCustom={onPickProfilePhoto}
+        t={t}
+      />
     </div>
   );
 }
